--- source/CarRentalManagement/src/CarRentalManagement.cpp
@@ -1,15 +1,0 @@
-//============================================================================
-// Name        : CarRentalManagement.cpp
-// Author      : 
-// Version     :
-// Copyright   : Your copyright notice
-// Description : Hello World in C++, Ansi-style
-//============================================================================
-
-#include <iostream>
-using namespace std;
-
-int main() {
-	cout << "Welcome to car rental management system" << endl; // prints Welcome to car rental management system
-	return 0;
-}
